massive overhaul to get going with express
Tweaks to routing and checking if user is logged in

remove keycloak.json

ignore keycloak.json

missing comma in launch debug config

debug with integrated terminal

hackish login checking

add route and remove console logging

the lobby will POST to / but then the redirect from the OIDC
will GET to / so a route needed to be added. I think this could
probably be consolidated to "app.use", but I'm not sure.

don't store the .env file

Removes the logout function of the web client and adds a quit button

rudimentary debug messages

quitbutton redirects browser to the app /quit endpoint

destroy the session when the user quits

this will force the user back to the lobby on a revisit after a quit

need to run bin/www now

change window redirection behavior

missed quit change on main menu